Refactor wallet showcase into separate component
Moves the wallet showcase logic and UI into a new `FirstWalletShowcase`
component. This cleans up the `WalletCards` component and prepares for
potential future variations of the wallet showcase.
